+++ src/main/java/jake/JakeException.java
@@ -1,3 +1,5 @@
+package jake;
+
 public class JakeException extends RuntimeException {
     public JakeException(String message) {
         super(message);
