+++ src/main/java/jake/Parser.java
@@ -1,3 +1,5 @@
+package jake;
+
 public class Parser {
 
     public static String getCommandWord(String fullCommand) {
