+++ src/main/java/jake/task/DeadlineTask.java
@@ -1,3 +1,7 @@
+package jake.task;
+
+import jake.JakeException;
+
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
