+++ src/main/java/jake/task/EventTask.java
@@ -1,3 +1,7 @@
+package jake.task;
+
+import jake.JakeException;
+
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
