+++ text-ui-test/EXPECTED.TXT
@@ -8,7 +8,7 @@ Hello from
 What can I do for you today?
 ____________________________________________________________
 ______________________________________________________________
-Todo task has been added:
+jake.task.Todo task has been added:
 [T][ ] golf swing
 Now you have 1 tasks in the list.
 ______________________________________________________________
